# Optimisation: replace three map-reduces with Renderers