--- a/render/container.go
+++ b/render/container.go
@@ -83,9 +83,7 @@ func (c connectionJoin) Render(rpt report.Report, dct Decorator) report.Nodes {
 		// Nodes without a hostid may be pseudo nodes - if so, pass through to result
 		if _, ok := m.Latest.Lookup(report.HostNodeID); !ok {
 			if id, ok := externalNodeID(m, addr, local); ok {
-				addToResults(m, id, ret, mapped, func() report.Node {
-					return report.MakeNode(id).WithTopology(Pseudo)
-				})
+				addToResults(m, id, ret, mapped, newPseudoNode)
 				continue
 			}
 		}
@@ -97,7 +95,7 @@ func (c connectionJoin) Render(rpt report.Report, dct Decorator) report.Nodes {
 			id, found = ipNodes[report.MakeScopedEndpointNodeID(scope, addr, port)]
 		}
 		if found && id != "" { // not one we blanked out earlier
-			addToResults(m, id, ret, mapped, func() report.Node {
+			addToResults(m, id, ret, mapped, func(id string) report.Node {
 				return inputNodes[id]
 			})
 		}