--- a/render/host.go
+++ b/render/host.go
@@ -104,6 +104,20 @@ func (e endpoints2Hosts) Render(rpt report.Report, dct Decorator) report.Nodes {
 	return ret
 }
 
+// Add Node M to the result set ret under id, creating a new result
+// node if not already there, and updating the old-id to new-id mapping
+// Note we do not update any counters for child topologies here
+func addToResults(m report.Node, id string, ret report.Nodes, mapped map[string]string, create func() report.Node) {
+	result, exists := ret[id]
+	if !exists {
+		result = create()
+	}
+	result.Children = result.Children.Add(m)
+	result.Children = result.Children.Merge(m.Children)
+	ret[result.ID] = result
+	mapped[m.ID] = result.ID
+}
+
 // Rewrite Adjacency for new nodes in ret, original nodes in input, and mapping old->new IDs in mapped
 func fixupAdjancencies(input, ret report.Nodes, mapped map[string]string) {
 	for _, n := range input {