--- a/render/id.go
+++ b/render/id.go
@@ -38,11 +38,20 @@ func NewDerivedPseudoNode(id string, node report.Node) report.Node {
 
 // NewDerivedExternalNode figures out if a node should be considered external and creates the corresponding pseudo node
 func NewDerivedExternalNode(n report.Node, addr string, local report.Networks) (report.Node, bool) {
+	id, ok := externalNodeID(n, addr, local)
+	if !ok {
+		return report.Node{}, false
+	}
+	return NewDerivedPseudoNode(id, n), true
+}
+
+// figure out if a node should be considered external and returns an ID which can be used to create a pseudo node
+func externalNodeID(n report.Node, addr string, local report.Networks) (string, bool) {
 	// First, check if it's a known service and emit a a specific node if it
 	// is. This needs to be done before checking IPs since known services can
 	// live in the same network, see https://github.com/weaveworks/scope/issues/2163
 	if hostname, found := DNSFirstMatch(n, isKnownService); found {
-		return NewDerivedPseudoNode(ServiceNodeIDPrefix+hostname, n), true
+		return ServiceNodeIDPrefix + hostname, true
 	}
 
 	// If the dstNodeAddr is not in a network local to this report, we emit an
@@ -50,13 +59,13 @@ func NewDerivedExternalNode(n report.Node, addr string, local report.Networks) (
 	if ip := net.ParseIP(addr); ip != nil && !local.Contains(ip) {
 		// emit one internet node for incoming, one for outgoing
 		if len(n.Adjacency) > 0 {
-			return NewDerivedPseudoNode(IncomingInternetID, n), true
+			return IncomingInternetID, true
 		}
-		return NewDerivedPseudoNode(OutgoingInternetID, n), true
+		return OutgoingInternetID, true
 	}
 
 	// The node is not external
-	return report.Node{}, false
+	return "", false
 }
 
 // DNSFirstMatch returns the first DNS name where match() returns