In three cases: `ConnectionJoin`, `ProcessRenderer` and `HostRenderer`, the code was running a `Map` to speculatively create `Node`s, then a `Reduce` to match them all up with something else.

If we replace those multiple operations with a single routine which does the matching directly, we save a lot of garbage.

I also added benchmark to time a single topology rendering, equivalent to hitting an endpoint like`/api/topology/hosts`. Best time of five runs; input is a 33MB report from production on Oct 15:
Before
```
$ go test -tags 'unsafe netgo' -run=x -bench=Topology -bench-report-file=prod-report1.json
BenchmarkTopologyList-2         	       1	1400508627 ns/op	352382920 B/op	 2913537 allocs/op
BenchmarkTopologyHosts-2        	       1	1124918587 ns/op	302509672 B/op	 2627937 allocs/op
BenchmarkTopologyContainers-2   	       2	 691307933 ns/op	194143244 B/op	 1723873 allocs/op
```

After
```
$ go test -tags 'unsafe netgo' -run=x -bench=Topology -bench-report-file=prod-report1.json
BenchmarkTopologyList-2         	       2	 692289010 ns/op	146370204 B/op	 1529315 allocs/op
BenchmarkTopologyHosts-2        	       3	 485628292 ns/op	95328261 B/op	 1245029 allocs/op
BenchmarkTopologyContainers-2   	      10	 208411667 ns/op	36021493 B/op	  515582 allocs/op
```

That is a 50-70% reduction in time and objects created.