@bboreham 2017-11-05
While working on this I had to write out the following analysis of the starting state; pasting it here for preservation:

```
TestShortLivedInternetNodeConnections calls
render.ContainerWithImageNameRenderer.Render(rpt, FilterNoop))

ContainerWithImageNameRenderer is a variable initialised with ContainerRenderer

First thing ContainerWithImageNameRenderer.Render() does is call
ContainerRenderer to get a list of "containers"

  ContainerRenderer is
    A merge of
      Map ColorConnectedProcessRenderer into MapProcess2Container
        ColorConnectedProcessRenderer calls ProcessRenderer then a function
          ProcessRenderer, if the report has any process nodes, is a merge of
            Map EndpointRenderer into MapEndpoint2Process
            SelectProcess
          the ColorConnected function adds a flag IsConnected to any nodes with edges
        MapProcess2Container creates container nodes if there is an ID, otherwise puts in 'uncontained'
      ConnectionJoin, which is a merge of
        Map a merge of
            Map r (SelectContainer) into nodeToIP via MapContainer2IP
            mapEndpoint2IP which is Map SelectEndpoint into endpoint2IP
          into ipToNode which removes any Endpoint2IP nodes that didn't match a container
        r (SelectContainer)
    Filtered to just nodes that have a ContainerState that is non-deleted (or pseudo-nodes)

Next Render() calls SelectContainerImage.Render(rpt, dct) to get a list of images

Then Render does:
  For each input node, if it has an image
    if the image is in the list of images
      if the image has a name
        updates the image details in the container object, and adds it to the output
  otherwise pass the input node straight through to the output
```