@bboreham 2017-11-05
How I tested it produces the same output as before:

First, obtain a test report, e.g. from your production system, and save it as a JSON file.

```
git checkout master
make prog/scope
prog/scope --mode=app --weave=false --app.collector=file:some-report.json &
curl -so /tmp/before http://localhost:4040/api/topology/containers
jq -S . < /tmp/before > /tmp/before-jq
```
(`-S` is used to sort keys so they are consistent)

then repeat on the branch for the 'after' case, and diff the `jq` output before and after.

Repeat for other topologies - `hosts`, etc.