@rade 2017-11-05
> Repeat for other topologies - `hosts`, etc.

I think `/api/topology` does that.