@bboreham 2017-11-06
All comments addressed