@rade 2017-11-06
Benchmarking this with a recent prod report by running
```
$ prog/scope --mode=app --weave=false --app.collector=file:prod-on-prod-report.json
```
and then timing topology rendering with
```
time curl -so /dev/null http://localhost:4040/api/topology/<topology>
```
yields

| topology   | before | after |
| ------------  | ------- | ------- |
| processes  | 0.307 | 0.252 |
| containers | 1.277 | 0.796 |
| pods       | 0.936 | 0.506 |
| hosts      | 1.227 | 0.711 |

So this is a significant improvement indeed.

LGTM. Merge it!